requirements.txt
alembic==1.5.5
astroid==2.5
BareNecessities==0.2.8
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-MDEditor==0.1.3
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==2.0.0rc1
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pylint==2.7.1
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3